requirements.txt
Django==2.0.13
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.19
django-nine==0.2.2
djangorestframework==3.10.2
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
shortuuid==0.5.0
six==1.12.0
urllib3==1.25.3
